+++ phpstan-bootstrap.php
@@ -7,7 +7,7 @@
 
 declare(strict_types=1);
 
-define( 'LW_IMG_VERSION', '1.5.0' );
+define( 'LW_IMG_VERSION', '1.6.0' );
 define( 'LW_IMG_FILE', __FILE__ );
 define( 'LW_IMG_PATH', __DIR__ . '/' );
 define( 'LW_IMG_URL', 'https://example.com/wp-content/plugins/lw-img/' );